Safari-609.1.11.1/LayoutTests/css3/flexbox/flexitem-stretch-image.html

<!DOCTYPE html>
<html>
<style>
.flexbox {
    width: 600px;
    display: flex;
    background-color: #aaa;
    position: relative;
    min-height: 10px;
}
.flexbox > * {
    margin: 0;
    border: 0;
    padding: 0;
    min-width: 0;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<div class="flexbox">
  <img data-expected-display="block" data-expected-width="345" style="flex: 1 0 auto;" src="../images/resources/blue-100.png">
  <img data-expected-display="block" data-expected-width="255" data-expected-height="100" style="flex: 1 0 auto;" src="../images/resources/green-10.png">
</div>

</body>
